+++ CHANGELOG.md
@@ -9,6 +9,8 @@ _Breaking change:_
 * Use `StateViewModel.listenState` instead of `ViewModel.listen`.
 * Use `ViewModel.listen` instead of `ViewModel.addListener`.
 
+* Support `ViewModel.read<T>` to read existing view model globally.
+
 ## 0.3.0
 
 * transfer to https://github.com/lwj1994/flutter_view_model. thank